app/src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
dataclasses==0.6
Flask==2.0.2
future==0.18.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yapf==0.31.0
zipp==3.6.0
